Simplify precisions in GLSL precision tests
The legacy_precision functions obscured the fact that almost all of the
precisions for 32-bit are the same with the float16_int8 and without.
Preserve the switch only where there was a functioal difference in the
old code.

Also, remove some parts that are dealing with lowp. They're never used,
so they're just making the code more complicated.

Components: Vulkan
Affects: dEQP-VK.glsl.builtin.precision*.exp.*
         dEQP-VK.glsl.builtin.precision*.exp2.*
         dEQP-VK.glsl.builtin.precision*.log.*
         dEQP-VK.glsl.builtin.precision*.log2.*
         dEQP-VK.glsl.builtin.precision*.sin.*
         dEQP-VK.glsl.builtin.precision*.cos.*
         dEQP-VK.glsl.builtin.precision*.acos.*
         dEQP-VK.glsl.builtin.precision*.atan.*

Change-Id: I1361ffb9ca7cd7327a2511f3b53ce2986c12a038