Fix image_extent test to handle OOM errors
The image_extent tests attempt to create a window with the maximum
supported extents, though it is not always possible to create such
a window due to platform memory constraints.

As long as the error was an out of memory error, allow the test to
pass. This fix only applies to the image_extent sub-test.

Affects:

dEQP-VK.wsi.*.swapchain.create.image_extent

Components: Vulkan

VK-GL-CTS issue: 1780

Change-Id: I5ac5439417424d101f6dc8d4f92da79b8fda2313