Always use GLSL version 450 for Vulkan testing
Since this isn't going to be consumed by the driver there's no
compatibility cost to just using high version for everything. Not having
the code for choosing the version is better for those reading and
maintaining the code and sets a better example for future code.

Components: Vulkan
Affects: dEQP-VK.glsl.texture_functions.*

Change-Id: I461d12b33acd7bc4039b09882ed98d9b9a46f49e