Don't allocate a GL_TEXTURE_2D_ARRAY with full render target width/height
On the following test:
KHR-GLES31.core.shader_image_load_store.advanced-sso-simple

Storage is specified as this:
glTexStorage3D(GL_TEXTURE_2D_ARRAY, 1, GL_RGBA32F, getWindowWidth(), getWindowHeight(), 8);

This is not an issue by default, as the default window is generall
small, but it is problematic when using the fbo surface type. In that
case the render target would be allocated with width and height as
GL_MAX_RENDERBUFFER_SIZE.

The problem is that most drivers doesn't support a texture with so big
dimensions.

This patch specify a default size, and small enough to be supported by
most driver, for the texture, as most tests on that group are already
doing.

Components: OpenGL-ES

VK-GL-CTS issue: 1905

Affects:
KHR-GLES31.core.shader_image_load_store.advanced-sso-simple

Change-Id: Ib998567151a230dd6b3550dbf6e014eed6cbb957