Combine ARB and Core subgroupBroadcast code
The ARB and core tests are attempting to do the same thing but using
slightly different names for the functions. Parameterise the names so
that we only have to maintain one copy of the test logic.

Also, make it so that we only have to write out the extensions that we
need once.

Components: Vulkan
Affects: dEQP-VK.subgroups.ballot_broadcast.*

Change-Id: I259249abba2b352dbd217aa4c26100ab7dbee637