Always require glslang and spirv-tools to be present
The prebuilt shader paths are no longer used. Just require both
dependencies to be present and drop the !DEQP_HAVE_GLSLANG /
!DEQP_HAVE_SPIRV_TOOLS paths.

Mention of prebuilt binaries is removed from the Vulkan CTS readme.

The machinery to actually build all the shaders upfront is left
in place for now as it has value as a presubmit check.

Components: Vulkan, OpenGL
VK-GL-CTS Issue: 1951

Change-Id: I6a79766a07e5aeb3c9c43236a3a220f8b654511e