Add list of tests for each fraction of CTS parallel run
The tests may be split in groups for parallel execution
using commandline deqp-fraction parameter.
This commit adds possibility to perform a list of tests
for each such fraction.
To achieve this - user must use parameter
deqp-fraction-mandatory-caselist-file defining a file with
a list of tests. File vk-fraction-mandatory-tests.txt
defines a list of mandatory tests when parallel execution
is applied.

Components: Vulkan Framework

VK-GL-CTS issue: 1897

Change-Id: I684eb07c24bbf3aeaf3386f205658dfe07b9fe25