Postpone .amber file parsing
All .amber files were parsed during test indexing, so they were parsed
even if none of the tests were executed. This wasn't a big performance
hit yet, but we're likely to increase the use of amber.

Due to some chicken-and-egg issues part of the capability check was
moved to the start of iterate, but since that throws InternalError if
the check fails this shouldn't be a performance issue.

The tests themselves have not changed, just the order of when to parse
the file.

Affects: None

Components: Vulkan, Framework

VK-GL-CTS issue: 899

Change-Id: Iacdfd9a61d439e3f5dec952b2426386a2eb3451c