Fix device group rendering CTS crash
dEQP-VK.device_group.* tests use --deqp-vk-device-group-id as the group
id to use and --deqp-vk-device-id as the index for the physical device
to use from that group when creating the logical device.

--deqp-vk-device-id is checked by the program, in general, to be valid
regarding the total number of physical devices in the system, but these
tests were not verifying it to be valid regarding the selected group.

In a machine with two devices in two separate groups (one device per
group), the program would accept the following options as valid, only to
segfault later:

  $ deqp-vk --deqp-vk-device-group-id=1 --deqp-vk-device-id=2 ...

  or

  $ deqp-vk --deqp-vk-device-group-id=2 --deqp-vk-device-id=2 ...

This commit checks the provided device id is valid for the selected
group.

Affected tests:
dEQP-VK.device_group.*

Components: Vulkan
VK-GL-CTS issue: 1553

Change-Id: Ied8f2fb73384303b793258cbaeaa7fad503a03c1