Prefer sized formats in internalformat tests
These tests create a static map of format/type pairs to reference
formats, and then don't update it in subsequent tests. This goes wrong
when the ES2 version of the tests run first and leave extension-only
unsized formats in the map in place of supported-by-ES3-core sized
formats. This change uses the sized format if the test is being run in
an ES3 context, even if the test itself is an ES2 test. This means that
the map will contain valid values if an ES3 test is subsequently run.

Also, change a few places to use the IsES3Compatible helper function
instead of reimplementing it every time.

Components: OpenGL ES
VK-GL-CTS issue: 1903
Affects: KHR-GLES*.core.internalformat.texture2d.*

Change-Id: Ia9a5248085dc8bdde1d459dd4245702e6e53d26e