Special-case RenderbufferStorageMultisampleEXT in codegen
Android has additional tests for this ES extension, which we had
observed was shipping broken on various Android devices. Unfortunately,
the ES extension which introduces this function specifies very different
semantics from the same-named function in desktop GL. In desktop GL,
this function is also defined to alias the core
RenderbufferStorageMultisample, and is assigned the same GLX opcode.

We had previously tried to unalias the function in the registry, but
that proved unworkable due to the GLX opcode aliasing.

Instead, add a special case in the CTS code generation.

Change-Id: Id62394cae2b2f01e0a91fdc4d9d357b634a26cf1
Components: Framework