Add OpFConvert precision tests
This commit adds tests to verify OpFConvert works as expected to convert
to and from 16, 32 and 64-bit wide floats in scalar and vector forms.

It does not use the existing precision tests framework due to the strict
precision requirements of OpFConvert, which require knowing the original
number exactly and are not suited for working with intervals.

New tests:
dEQP-VK.glsl.builtin.precision_fconvert.*

Components: Vulkan, Framework
VK-GL-CTS issue: 1868

Change-Id: I4161dae9d9f2c494cc873d5d11b40286bdfb6466