Make texture aniso basic self tests less sensitive
The tests cases first check with a fuzzy diff that the images are more
or less the same.

The second diff then checks with a standard image comparison + tolerance
that the images are not identical.

Some of IMG's implementations produce a closer image than the test is
currently expecting - we need to lower the tolerance from 0.05 to 0.02
in order for the diff algorithm to detect that the images are
different 'enough'.

Though there is no actual spec requirement, it is 'generally going to
be the case', for any implementation, that the images will have some
amount of difference. It is potentially useful to still have this test
inplace to check againsty aniso being accidently disabled in the tested
implementation, rather than remove it altogether.

Affects:
	dEQP-VK.texture.filtering_anisotropy.basic.*

Components: Vulkan

VK-GL-CTS Issue: 2047

Change-Id: Ia9148e12172ac1c4757f5007e57b3a2b3a3434eb