Avoid undefined behaviour when calculating subgroup ballot masks
The affected tests are shifting the bits of the last component of the
computed ballot masks by 32. However, the SPIR-V specification
establishes that the result of OpShift* is undefined if `Shift` is
greater than *or equal* to the bit width of the components of `Base`.

Affects:

dEQP-VK.subgroups.ballot_other.*

Components: Vulkan

VK-GL-CTS issue: 2083

Change-Id: I8c9b4373f1429669a40741986916e8d71b7eb573