Simplify the test code for ballot bit counts
The old code was quite obscure and had many references to 'Undef'
despite nothing really being undefined. The 'undefTerritory' that was
referred to was actually places where the result should be 0.

The test was also doing one extra loop iteration (i == 128) which
duplicated the prior test for 'allZeros'.

This change highlights the fact that there is no testing that these are
returning the correct results. They are only ever checking whether the
result is 0 or not.

Components: Vulkan
VK-GL-CTS issue: 2192
Affects: dEQP-VK.subgroups.ballot_other.*bitcount*

Change-Id: I0e1f91b60f103311d0b75748ff21ac6005e01cfd