Fix use of pointer to stack allocated struct
The dEQP-VK.wsi.display_control.swapchain_counter test's
createSwapchainConfig was returning a struct containing a pointer to
another struct that was stack allocated, so after the return that
pointer is no longer valid.

This change stores the inner struct in the SwapchainCounterTestInstance
instead of on the stack of createSwapchainConfig, and uses a pointer to
that struct, which remains valid while the test runs.

Components: Vulkan

VK-GL-CTS issue: 2184

Affects:
dEQP-VK.wsi.display_control.swapchain_counter

Change-Id: I8e83f810427167113e5cee843cbafda237606dcf