Stricter codomain for atan tests
If the codomain is going to be enforced strictly then we should update
the precision tests to check it. The atan precision test looks to have
had an issue applying the codomain incorrectly which meant that large
values of the arguments were not tested at all.

Components: Vulkan
VK-GL-CTS issue: 479
Affects: dEQP-VK.glsl.builtin.precision*atan.*

Change-Id: I842adb035cb59fe8ff1423f11acdad0a90ee1549