Validate samplerless descriptor writes
This commits adds a few tests that update image-related descriptors for
descriptor types in which the sampler field of the VkDescriptorImageInfo
structure, used in VkWriteDescriptorSet, is supposed to be ignored.
Given that premise, the tests try to set the sampler field to different
values like VK_NULL_HANDLE, a fixed value of 1 and the handler for a
destroyed sampler. This should have no side effects.

New tests:
dEQP-VK.binding_model.descriptor_update.samplerless.*

Components: Vulkan
VK-GL-CTS issue: 2172

Change-Id: I87bf4226738ecd2ff7d2c7aa23551f22d43a8637