Test exact sampling cases
This commit adds test to check samplers do not wrongly manipulate the
value of sampled texels from a texture.

This is checked by creating a texture with a single mipmap level and a
framebuffer image of the same size. Then, a full-screen quad is drawn
with appropriate vertex and texture coordinates so it covers the whole
screen. When drawing, the sampler uses the nearest filter, which means
the output framebuffer image and the texture should match exactly.

New tests:
dEQP-VK.pipeline.sampler.exact_sampling.*

Components: Vulkan
VK-GL-CTS issue: 2170

Change-Id: Ia5fb27d577d7bca92cbc723c798f8ab06190453b