Check if display is available for Vulkan.
If the display is already being used by native windowing system,
then it may not be available to Vulkan, which may cause the test
to fail.
To account for this, check if any windowing system is already present
and running and then report the test to be unsupported.
Since the extension is meant to support direct rendering to display,
avoid running it when a windowing system is present.

Components: Vulkan

VK-GL-CTS issue: #2230

Affects:
dEQP-VK.wsi.display_control.swapchain_counter

Change-Id: I7fd00b417021e4a2ef017b742e529f5a3917a056
(cherry picked from commit 5318e8023d181def3d6affe446ab251f9898737e)