Add tests for usage of pixel storage modes
Pixel storage modes allow to select a subregion of data when
specifying a texture image.

For example, when specifying a 2D texture image, by setting both
GL_UNPACK_SKIP_PIXELS and GL_UNPACK_SKIP_ROWS, a subrectangle can be
selected.

This set of tests contains subcases for:

- glTexImage2D
- glTexImage3D
- glCompressedTexImage2D
- glCompressedTexImage3D

Source data is filled with black colour with a white (or different
shades of grey for each layer in case of 3D textures) subregion
somewhere within it. This subregion is then selected by setting pixel
storage modes and used as a source when specifying a texture. The
correct colour is checked in a shader.

Compressed cases use two compression formats:

- GL_COMPRESSED_RGB_S3TC_DXT1_EXT
- GL_COMPRESSED_RGBA_ASTC_8x5_KHR

Adds a new subgroup of tests to two groups:

- KHR-GL33.pixelstoragemodes.*
- KHR-GLES31.core.pixelstoragemodes.*

Signed-off-by: Yevhenii Kolesnikov <yevhenii.kolesnikov@globallogic.com>