OpenGL ES CTS fixes for passing on WGL
When running cts-runner --type=es32 on a WGL platform it will pick up several
pixel formats for the default framebuffer that aren't supported correctly
by the existing OpenGL ES 3.2 tests.

1. OpenGL ES 3.2 CTS expects the platform to expose rgb565, but that is not
available on WGL and should not be a requirement of OpenGL ES 3.2
conformance. This CL disables rgb565 is a requirement for all non-Android
platforms, which extends the change originally made for X11 in
https://gerrit.khronos.org/#/c/4615/ for issue VK-GL-CTS issue: 1474.

2. The 10-bit+ bits per pixel component WGL pixel formats are skipped since
the existing OpenGL ES 3.2 tests, which use pixel transfer functions
like glReadPixels, and assume an 8-bit type, don't work correctly since
glReadPixels is not defined to do format and type conversions like it
does with OpenGL.

3. Default framebuffers with sRGB color space are skipped because
under OpenGL ES sRGB encoded will default for on in this case and the
existing OpenGL ES tests do not currently support a non-linear
default framebuffer. This is not a problem with OpenGL since with
OpenGL sRGB encoding has to be enabled explicitly.

Affects: cts-runner --type=es32 on WGL

Components: OpenGL

VK-GL-CTS issue: 2257

Change-Id: If12aff8c10571b7c0f72da618b13753c340a60c4