Avoid exceeding maxPerStageResources in descriptor_indexing tests
The 'calculateInLoop; (*_in_loop) variant of the test is using additional
uniform texel buffer descriptor binding in pipeline layout.

This binding was counted only against maxPerStageDescriptorSampledImages and
maxDescriptorSetSampledImages limits, but not maxPerStageResources.

The fix is to count it against all three limits - but do this only when needed (in
*_in_loop) tests.

Component: Vulkan

VK-GL-CTS Issue: 2333

Affects: dEQP-VK.descriptor_indexing.*

Change-Id: I1c3476cb8c50c3ed61e73ecd4c106e6e9b613022