Fix broken layout binding tests
Some of the shaders being generated would fail to compile due
to a misplaced ;

An example of the code being gererated is as follows:

  fragColor = vec4(float(atomicCounter(atomic0)), 1.0, 0.0, 1.0);
  +vec4(float(atomicCounter(atomic1)), 1.0, 0.0, 1.0);
  +vec4(float(atomicCounter(atomic2)), 1.0, 0.0, 1.0);
  +vec4(float(atomicCounter(atomic3)), 1.0, 0.0, 1.0);

So we had compile tests checking for failures and passing but for
the wrong reason.

Some of the offset checks also had broken shaders being generated.