Clarify integer formats conversions in blit_image tests
This change just updates comments to make code clear.
It was not clear if getTextureChannelClass identified how values are
stored or how they are sampled but it was specified in description of
TextureChannelClass. This change clarifies description of
getTextureChannelClass and comments in blit tests to avoid confusion
in future.

Components: Vulkan

VK-GL-CTS public issue: 174

Change-Id: I49944508337d2950eb8066bf8e3b9ff3db855668