Flip winding of one triangle in watertightness test
This flips the winding order of one of two triangles
in watertightness test, so they do share an edge. According
to spec, the triangles share and edge only if they are in
the same winding.

Spec says:
"""
For a set of triangles with identical transforms, within a single instance:
...
  * Any set of two triangles with two shared vertices that were specified
    in the same winding order in each triangle have a shared edge defined
    by those vertices.
...
Implementations should not double-hit or miss when a ray intersects
a shared edge, or a shared vertex of a closed fan.
"""

Components: Vulkan

VK-GL-CTS Issue: 2409

Change-Id: Ia3a277c3e157da6c312e8ff180275763d61ee33e
Affects: dEQP-VK.ray_tracing.watertightness.*