Linux WSI checks the WS availability on create
Before creating a display of a specific WSI type,
check for the server availability, first. If the display server of
the requested WSI type (XLIB/XCB/WAYLAND) is not present/running,
then when throw an exception of type NotSupportedError
rather than one of type ResourceError.

This would allow for the tests to be marked as unsupported,
rather than causing the CTS to crash/exit.

Affects: dEQP-VK.wsi.wayland.*
         dEQP-VK.wsi.xcb.*
         dEQP-VK.wsi.xlib.*
         dEQP-VK.image.swapchain*

Components: Vulkan, WSI, Linux platform

VK-GL-CTS issue: 2422

Change-Id: I7538be79e528ed9cde38cff3404b5f3603bdf5c1