Robustness2 features are enabled by default
Robustness2 features are enabled by default for the Default Device,
but robustBufferAccess is disabled. The correct behaviour is to not
enable robustness2 features for this device and allow robustness2
tests to create a separate device with the features turned on.

Components: Framework

VK-GL-CTS issue: 2408

Change-Id: I5b5afd24d497340bd48d97c86b5cb17c509f1d72
(cherry picked from commit a20baf416c7dd115bdf3307fb9363bdabbb821ce)