Experimental crash postmortem use-after-free test
Adds a tests to deliberately free a buffer before executing a compute
shader that accesses it for the purposes of deliberately triggering
device-loss.

This test is intended for experimental/investigative use by the crash
postmortem TSG, is expected to fail, and should not be included in
mustpass lists.

Added Tests:

dEQP-VK-experimental.postmortem.use_after_free.*

Components: Vulkan

Vulkan issue: 2255

Change-Id: Icb7f47a37e864932044343493057ef4882473b94