Experimental crash postmortem shader timeout test
Adds a set of tests to execute a compute shader containing large
nested loops for the purposes of deliberately triggering timeouts and
device-loss.

This test is intended for experimental/investigative use by the crash
postmortem TSG, is expected to fail, and should not be included in
mustpass lists.

Added Tests:

dEQP-VK-experimental.postmortem.shader_timeout.*

Components: Vulkan

Vulkan issue: 2255

Change-Id: Ie4272d48d523c50a18c337447b16c36df7bd9bed