Check large instanceCustomIndex values
Add a new set of tests that check the value of the
InstanceCustomIndexKHR builtin from the closest-hit, intersection and
any-hit shaders correctly matches the instanceCustomIndex value when
creating instances. The tests use large values for instanceCustomIndex,
making sure the most-significant bit when represented as a 24-bit value
is a 1 and it's not incorrectly sign-extended by the implementation when
presented as the InstanceCustomIndexKHR builtin in shaders.

This value is combined with an all-ones mask in the highest bits of the
32-bit dword.

New tests:
dEQP-VK.ray_tracing_pipeline.acceleration_structures.instance_index.*

Affected tests:
dEQP-VK.ray_tracing_pipeline.acceleration_structures.*

Components: Vulkan
VK-GL-CTS issue: 2586

Change-Id: I2af85c9e2a2c78075502ab42db451651c21e0f1a